fix: rename folder to keep convention
