+++ src/components/inscription-preview/index.js
